JWT token fix

+++ controller/users.js
@@ -171,21 +171,30 @@ const upload = multer({ storage: storage });
 const avatars = async (req, res, next) => {
   const { path: temporaryName, originalname } = req.file;
   const filename = path.join(uploadDir, originalname);
-  const { email } = req.user;
+  const { user } = req;
+  const { email, token } = user;
   const username = email.split("@")[0];
   const newAvatarPath = `${storeImage}/${username}.jpg`;
 
   try {
+    if (!token) return res.status(401).json({ message: "Not authorized" });
+
     await fs.rename(temporaryName, filename);
 
     const avatarImg = await Jimp.read(filename);
     avatarImg.resize(250, 250).write(newAvatarPath);
 
+    user.avatarURL = newAvatarPath;
+    await user.save();
+
     await fs.unlink(filename);
 
-    return res.status(200).json({ avatarURL: newAvatarPath });
+    const { avatarURL } = user;
+
+    return res.status(200).json({ avatarURL });
   } catch (error) {
     await fs.unlink(temporaryName);
+    console.log(error);
     return res.status(401).json({ message: "Not authorized" });
   }
 };
